requirements.txt
Flask==2.0.3
Pillow==9.2.0
tensorflow==2.9.1
gunicorn==20.1.0
